Show StopWatch on the Lock-Screen
Sets the Visibility of the Notification to VISIBILITY_PUBLIC instead of
using the default VISIBILITY_PRIVATE.

My main interest in this app is to be able to stop a stopwatch time
easily w/o having to unlock my phone. I don't know of a use case in
which this would be a privacy issue.